refactor(ramp): A4 extract properties-generators.mjs
Split get-type-spec-defaults-map.mjs (588 lines, misnamed) into two files:

properties-generators.mjs — the properties generator pipeline:
  fontGen, baseFontSizeGen, fontSizeGen (font/size mini-pipeline)
  axisLocationsGen, openTypeFeaturesGen, languageTagGen (leaf generators)
  createMargin, marginsGen (margin pipeline)
  AutoLinearLeadingSyntheticValue, leadingGen (leading pipeline)
  calculateFontAxisValueSynthetic, axesMathAxisLocationValueGen,
  axisMathLocationsGen (axes-math pipeline)
  REGISTERED_GENERIC_TYPESPEC_FIELDS, TYPE_SPEC_PROPERTIES_GENERATORS,
  STYLE_PATCH_PROPERTIES_GENERATORS (registries — trunk)

get-type-spec-defaults-map.mjs — now contains only the defaults tail:
  _skipPrefix, _skipFullKey, _getTypeSpecDefaultsMap (101 lines)

Updated consumers:
  style-patch-source-live-properties.typeroof.jsx → properties-generators.mjs
  index.typeroof.jsx → properties-generators.mjs (TYPE_SPEC_PROPERTIES_GENERATORS)

npm run typecheck: clean.

model: anthropic/claude-sonnet-4.6
provider: openrouter
agent: goose v1.31.0
