+++ lib/js/components/layouts/ramp/style-patch-source-live-properties.typeroof.jsx
@@ -1,5 +1,5 @@
 import { _BaseComponent } from "../../basics.mjs";
-import { STYLE_PATCH_PROPERTIES_GENERATORS } from "./get-type-spec-defaults-map.mjs";
+import { STYLE_PATCH_PROPERTIES_GENERATORS } from "./properties-generators.mjs";
 
 /**
  * This is to determine and cache the set of properties of a StylePatch.
